patch 8.1.2252: compiler warning for int size
Problem:    Compiler warning for int size.
Solution:   Add type cast. (Mike Williams)

+++ src/filepath.c
@@ -79,7 +79,7 @@ get_short_pathname(char_u **fnamep, char_u **bufp, int *fnamelen)
     vim_free(wfname);
     vim_free(newbuf);
 
-    *fnamelen = l == 0 ? l : STRLEN(*bufp);
+    *fnamelen = l == 0 ? l : (int)STRLEN(*bufp);
     return OK;
 }
